move ProcessGrid into a stand-alone header under Utilities::MPI namespace
