+++ include/deal.II/lac/scalapack.h
@@ -21,6 +21,7 @@
 #ifdef DEAL_II_WITH_SCALAPACK
 
 #include <deal.II/base/exceptions.h>
+#include <deal.II/base/process_grid.h>
 #include <deal.II/lac/full_matrix.h>
 #include <deal.II/lac/lapack_support.h>
 #include <deal.II/base/mpi.h>
@@ -30,168 +31,6 @@
 #include <memory>
 
 DEAL_II_NAMESPACE_OPEN
-
-
-// Forward declaration of class ScaLAPACKMatrix for ProcessGrid
-template <typename NumberType> class ScaLAPACKMatrix;
-
-
-/**
- * A class taking care of setting up a two-dimensional processor grid.
- * For example an MPI communicator with 5 processes can be arranged into a
- * 2x2 grid with the 5-th processor being inactive:
- * @code
- *      |   0     |   1
- * -----| ------- |-----
- * 0    |   P0    |  P1
- *      |         |
- * -----| ------- |-----
- * 1    |   P2    |  P3
- * @endcode
- *
- * A shared pointer to this class is provided to ScaLAPACKMatrix matrices to
- * perform block-cyclic distribution.
- *
- * Note that this class allows to setup a process grid which has fewer
- * MPI cores than the total number of cores in the communicator.
- *
- * Currently the only place where one would use a ProcessGrid object is
- * in connection with a ScaLAPACKMatrix object.
- *
- * @author Benjamin Brands, 2017
- */
-class ProcessGrid
-{
-public:
-
-  /**
-   * Declare class ScaLAPACK as friend to provide access to private members.
-   */
-  template <typename NumberType> friend class ScaLAPACKMatrix;
-
-  /**
-   * Constructor for a process grid with @p n_rows and @p n_columns for a given @p mpi_communicator.
-   * The product of rows and columns should be less or equal to the total number of cores
-   * in the @p mpi_communicator.
-   */
-  ProcessGrid(MPI_Comm mpi_communicator,
-              const unsigned int n_rows,
-              const unsigned int n_columns);
-
-  /**
-   * Constructor for a process grid for a given @p mpi_communicator.
-   * In this case the process grid is heuristically chosen based on the
-   * dimensions and block-cyclic distribution of a target matrix provided
-   * in @p n_rows_matrix, @p n_columns_matrix, @p row_block_size and @p column_block_size.
-   *
-   * The maximum number of MPI cores one can utilize is $\min\{\frac{M}{MB}\frac{N}{NB}, Np\}$, where $M,N$
-   * are the matrix dimension and $MB,NB$ are the block sizes and $Np$ is the number of
-   * processes in the @p mpi_communicator. This function then creates a 2D processor grid
-   * assuming the ratio between number of process row $p$ and columns $q$ to be
-   * equal the ratio between matrix dimensions $M$ and $N$.
-   *
-   * For example, a square matrix $640x640$ with the block size $32$
-   * and the @p mpi_communicator with 11 cores will result in the $3x3$
-   * process grid.
-   */
-  ProcessGrid(MPI_Comm mpi_communicator,
-              const unsigned int n_rows_matrix,
-              const unsigned int n_columns_matrix,
-              const unsigned int row_block_size,
-              const unsigned int column_block_size);
-
-  /**
-  * Destructor.
-  */
-  ~ProcessGrid();
-
-  /**
-  * Return the number of rows in the processes grid.
-  */
-  unsigned int get_process_grid_rows() const;
-
-  /**
-  * Return the number of columns in the processes grid.
-  */
-  unsigned int get_process_grid_columns() const;
-
-  /**
-   * Send @p count values stored consequently starting at @p value from
-   * the process with rank zero to processes which
-   * are not in the process grid.
-   */
-  template <typename NumberType>
-  void send_to_inactive(NumberType *value, const int count=1) const;
-
-  /**
-   * Return <code>true</code> if the process is active within the grid.
-   */
-  bool is_process_active() const;
-
-private:
-
-  /**
-   * A private constructor which takes grid dimensions as an <code>std::pair</code>.
-   */
-  ProcessGrid(MPI_Comm mpi_communicator,
-              const std::pair<unsigned int,unsigned int> &grid_dimensions);
-
-  /**
-  * An MPI communicator with all processes (active and inactive).
-  */
-  MPI_Comm mpi_communicator;
-
-  /**
-  * An MPI communicator with inactive processes and the process with rank zero.
-  */
-  MPI_Comm mpi_communicator_inactive_with_root;
-
-  /**
-  * BLACS context. This is equivalent to MPI communicators and is
-  * used by ScaLAPACK.
-  */
-  int blacs_context;
-
-  /**
-  * Rank of this MPI process.
-  */
-  const unsigned int this_mpi_process;
-
-  /**
-  * Total number of MPI processes.
-  */
-  const unsigned int n_mpi_processes;
-
-  /**
-  * Number of rows in the processes grid.
-  */
-  int n_process_rows;
-
-  /**
-  * Number of columns in the processes grid.
-  */
-  int n_process_columns;
-
-  /**
-  * Row of this process in the grid.
-  *
-  * It's negative for in-active processes.
-  */
-  int this_process_row;
-
-  /**
-  * Column of this process in the grid.
-  *
-  * It's negative for in-active processes.
-  */
-  int this_process_column;
-
-  /**
-  * A flag which is true for processes within the 2D process grid.
-  */
-  bool mpi_process_is_active;
-};
-
 
 /**
  * A wrapper class around ScaLAPACK parallel dense linear algebra.
@@ -278,7 +117,7 @@ public:
    */
   ScaLAPACKMatrix(const size_type n_rows,
                   const size_type n_columns,
-                  const std::shared_ptr<const ProcessGrid> process_grid,
+                  const std::shared_ptr<const Utilities::MPI::ProcessGrid> process_grid,
                   const size_type row_block_size = 32,
                   const size_type column_block_size = 32,
                   const LAPACKSupport::Property property = LAPACKSupport::Property::general);
@@ -288,7 +127,7 @@ public:
    * using the process grid in @p process_grid.
    */
   ScaLAPACKMatrix(const size_type size,
-                  const std::shared_ptr<const ProcessGrid> process_grid,
+                  const std::shared_ptr<const Utilities::MPI::ProcessGrid> process_grid,
                   const size_type block_size = 32,
                   const LAPACKSupport::Property property = LAPACKSupport::Property::symmetric);
 
@@ -439,10 +278,10 @@ private:
   LAPACKSupport::Property property;
 
   /**
-   * A shared pointer to a ProcessGrid object which contains a BLACS context
+   * A shared pointer to a Utilities::MPI::ProcessGrid object which contains a BLACS context
    * and a MPI communicator, as well as other necessary data structures.
    */
-  std::shared_ptr<const ProcessGrid> grid;
+  std::shared_ptr<const Utilities::MPI::ProcessGrid> grid;
 
   /**
    * Number of rows in the matrix.
@@ -527,6 +366,8 @@ private:
 
 // ----------------------- inline functions ----------------------------
 
+#ifndef DOXYGEN
+
 template <typename NumberType>
 inline
 NumberType
@@ -544,6 +385,8 @@ ScaLAPACKMatrix<NumberType>::local_el(const int loc_row, const int loc_column)
 {
   return (*this)(loc_row,loc_column);
 }
+
+#endif // DOXYGEN
 
 DEAL_II_NAMESPACE_CLOSE
