+++ source/base/CMakeLists.txt
@@ -64,6 +64,7 @@ SET(_unity_include_src
   polynomials_piecewise.cc
   polynomials_rannacher_turek.cc
   polynomials_raviart_thomas.cc
+  process_grid.cc
   quadrature.cc
   quadrature_lib.cc
   quadrature_selector.cc
